rakcha: feat/upgrade + security history scrub

+++ apps/web/src/Controller/SeriesController.php
@@ -7,14 +7,11 @@ use App\Entity\Series;
 use App\Form\SeriesType;
 use App\Repository\SeriesRepository;
 use Doctrine\ORM\EntityManagerInterface;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use Twilio\Rest\Client;
-
 
 #[Route('/series')]
 class SeriesController extends AbstractController
@@ -22,12 +19,14 @@ class SeriesController extends AbstractController
     #[Route('/', name: 'app_series_index', methods: ['GET'])]
     public function index(SeriesRepository $seriesRepository, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
         $statisticsData = $seriesRepository->getStatisticsByCategory();
         $form = $this->createForm(SeriesType::class, new Series());
-        $updateForms = array();
-        for ($i = 0; $i < count($seriesRepository->findAll()); $i++) {
+        $updateForms = [];
+        for ($i = 0; $i < count($seriesRepository->findAll()); ++$i) {
             $updateForms[$i] = $this->createForm(SeriesType::class, $seriesRepository->findAll()[$i])->createView();
         }
+
         return $this->render('back/seriesTables.html.twig', [
             'statisticsData' => $statisticsData,
             'series' => $seriesRepository->findAll(),
@@ -42,9 +41,9 @@ class SeriesController extends AbstractController
         /*$searchTerm = $request->query->get('nom');
         $sortBy = $request->query->get('sort_by', 'nom'); // Tri par défaut par 'nom'
         $sortOrder = $request->query->get('sort_order', 'asc'); // Ordre de tri par défaut 'asc'
-        
+
         $queryBuilder = $entityManager->getRepository(Series::class)->createQueryBuilder('s');
-        
+
         // Requête de recherche
         if ($searchTerm) {
             $queryBuilder->where('s.nom LIKE :searchTerm')
@@ -52,10 +51,10 @@ class SeriesController extends AbstractController
                          ->orWhere('s.pays LIKE :searchTerm')
                          ->setParameter('searchTerm', '%' . $searchTerm . '%');
         }
-        
+
         // Requête de tri
         $queryBuilder->orderBy('s.' . $sortBy, $sortOrder);
-        
+
         $series = $queryBuilder->getQuery()->getResult();
     */
 
@@ -70,12 +69,12 @@ class SeriesController extends AbstractController
             ->getResult();
         /*
             // Récupérer l'identifiant de la catégorie de comédie
-            $comedyCategoryId = 4; 
+            $comedyCategoryId = 4;
             // Récupérer les recommandations de séries
             $recommendations = $seriesRepository->findRelaxingSeries($comedyCategoryId);
         */
         // Créer un formulaire pour chaque série
-        $updateForms = array();
+        $updateForms = [];
         foreach ($allSeries as $serie) {
             $updateForms[] = $this->createForm(SeriesType::class, $serie)->createView();
         }
@@ -93,6 +92,8 @@ class SeriesController extends AbstractController
     #[Route('/new', name: 'app_series_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $series = new Series();
         $form = $this->createForm(SeriesType::class, $series);
         $form->handleRequest($request);
@@ -106,43 +107,22 @@ class SeriesController extends AbstractController
                     // extension cannot be guessed
                     $extension = 'bin';
                 }
-                $filename = rand(1, 99999) . '.' . $extension;
-                $destination = $this->getParameter('kernel.project_dir') . "/public/img/series";
+                $filename = rand(1, 99999).'.'.$extension;
+                $destination = $this->getParameter('kernel.project_dir').'/public/img/series';
                 $file->move($destination, $filename);
-                $series->setImage("/img/series/" . $filename);
+                $series->setImage('/img/series/'.$filename);
 
                 // Copy the file to another location
-                $anotherDestination = "C:\\xampp\\htdocs\\Rakcha\\rakcha-desktop\\src\\main\\resources\\img\\series";
-                copy($destination . "/" . $filename, $anotherDestination . "/" . $filename);
+                $anotherDestination = 'C:\\xampp\\htdocs\\Rakcha\\rakcha-desktop\\src\\main\\resources\\img\\series';
+                copy($destination.'/'.$filename, $anotherDestination.'/'.$filename);
             }
             $entityManager->persist($series);
             $entityManager->flush();
-            /*
-           // Envoi du SMS après l'ajout de la série
-        $twilioSid = "ACd3d2094ef7f546619e892605940f1631";
-        $twilioToken = "8d56f8a04d84ff2393de4ea888f677a1";
-        $twilioPhoneNumber = "+17573640849";
-        $phoneNumber = '+21653775010'; // Remplacez par le numéro de téléphone réel de votre base de données
-
-      try {
-          $client = new Client($twilioSid, $twilioToken);
-          $client->messages->create(
-              $phoneNumber,
-              [
-                  'from' => $twilioPhoneNumber,
-                  'body' => 'New Serie is Out There'
-              ]
-          );
-      } catch (\Exception $e) {
-          // Gérer l'exception ici
-          $errorMessage = $e->getMessage();
-          $this->addFlash('error', 'Erreur lors de l\'envoi du SMS : ' . $errorMessage);
-      }
-        */
 
             return $this->redirectToRoute('app_series_index', [], Response::HTTP_SEE_OTHER);
         }
         dump($form->getErrors(true));
+
         return $this->renderForm('series/new.html.twig', [
             'series' => $series,
             'form' => $form,
@@ -160,6 +140,8 @@ class SeriesController extends AbstractController
     #[Route('/{idserie}/edit', name: 'app_series_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Series $series, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $form = $this->createForm(SeriesType::class, $series);
         $form->handleRequest($request);
 
@@ -172,14 +154,14 @@ class SeriesController extends AbstractController
                     // extension cannot be guessed
                     $extension = 'bin';
                 }
-                $filename = rand(1, 99999) . '.' . $extension;
-                $destination = $this->getParameter('kernel.project_dir') . "/public/img/series";
+                $filename = rand(1, 99999).'.'.$extension;
+                $destination = $this->getParameter('kernel.project_dir').'/public/img/series';
                 $file->move($destination, $filename);
-                $series->setImage("/img/series/" . $filename);
+                $series->setImage('/img/series/'.$filename);
 
                 // Copy the file to another location
-                $anotherDestination = "C:\\xampp\\htdocs\\Rakcha\\rakcha-desktop\\src\\main\\resources\\img\\series";
-                copy($destination . "/" . $filename, $anotherDestination . "/" . $filename);
+                $anotherDestination = 'C:\\xampp\\htdocs\\Rakcha\\rakcha-desktop\\src\\main\\resources\\img\\series';
+                copy($destination.'/'.$filename, $anotherDestination.'/'.$filename);
             }
             $entityManager->flush();
 
@@ -195,7 +177,9 @@ class SeriesController extends AbstractController
     #[Route('/{idserie}', name: 'app_series_delete', methods: ['POST'])]
     public function delete(Request $request, Series $series, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $series->getIdserie(), $request->request->get('_token'))) {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+        if ($this->isCsrfTokenValid('delete'.$series->getIdserie(), $request->request->get('_token'))) {
             $entityManager->remove($series);
             $entityManager->flush();
         }
@@ -206,6 +190,8 @@ class SeriesController extends AbstractController
     #[Route('/series/{idserie}/like', name: 'app_like_series', methods: ['GET'])]
     public function likeSeries(int $idserie, EntityManagerInterface $entityManager): RedirectResponse
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_REMEMBERED');
+
         // Récupérer la série depuis la base de données en fonction de l'ID
         $series = $entityManager->getRepository(Series::class)->find($idserie);
 
@@ -220,10 +206,11 @@ class SeriesController extends AbstractController
         return new RedirectResponse($this->generateUrl('app_series_liste'));
     }
 
-
     #[Route('/series/{idserie}/dislike', name: 'app_dislike_series', methods: ['GET'])]
     public function dislikeSeries(int $idserie, EntityManagerInterface $entityManager): RedirectResponse
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_REMEMBERED');
+
         // Récupérer la série depuis la base de données en fonction de l'ID
         $series = $entityManager->getRepository(Series::class)->find($idserie);
 
@@ -237,13 +224,6 @@ class SeriesController extends AbstractController
         // Rediriger l'utilisateur
         return $this->redirectToRoute('app_series_liste');
     }
-
-
-
-
-
-
-
 
     /*
     #[Route('/{idSerie}/toggle-favorite', name: 'app_toggle_favorite', methods: ['POST'])]
@@ -267,9 +247,9 @@ class SeriesController extends AbstractController
             $favoris->setIdSerie($idSerie);
             $entityManager->persist($favoris);
         }
-        
+
         $entityManager->flush();
-        
+
         // Rediriger vers la liste des favoris
         return $this->redirectToRoute('app_favorites_list');
     }
@@ -316,6 +296,6 @@ class SeriesController extends AbstractController
 
             return $response;
         }
-       
+
     */
 }
